Keep peer manager tests under size gate
