+++ crates/oasis7_net/src/libp2p_net/peer_manager.rs
@@ -456,28 +456,20 @@ mod tests {
         peer_id: PeerId,
         discovery_sources: Vec<PeerDiscoverySource>,
     ) -> SignedPeerRecord {
-        sample_record_with_role(peer_id, PeerNodeRole::FullStorage, discovery_sources)
-    }
-
-    fn sample_record_with_role(
-        peer_id: PeerId,
-        node_role: PeerNodeRole,
-        discovery_sources: Vec<PeerDiscoverySource>,
-    ) -> SignedPeerRecord {
         SignedPeerRecord {
             record: oasis7_proto::distributed_dht::PeerRecord {
                 peer_id: peer_id.to_string(),
                 node_id: format!("node-{peer_id}"),
                 world_id: "world-a".to_string(),
                 network_id: "world-a".to_string(),
-                node_role: node_role.as_str().to_string(),
+                node_role: PeerNodeRole::FullStorage.as_str().to_string(),
                 deployment_mode: PeerDeploymentMode::Hybrid,
                 reachability_class: PeerReachabilityClass::Hybrid,
                 direct_addrs: Vec::new(),
                 hole_punch_addrs: Vec::new(),
                 relay_addrs: Vec::new(),
                 discovery_sources,
-                capability_lanes: node_role.default_capability_lanes(),
+                capability_lanes: PeerNodeRole::FullStorage.default_capability_lanes(),
                 source_operator: None,
                 source_asn: None,
                 published_at_ms: 0,
@@ -664,146 +656,6 @@ mod tests {
         for peer in [peer_a, peer_b, peer_c] {
             assert_eq!(healths[&peer].status, PeerManagerHealthStatus::Active);
         }
-    }
-
-    #[test]
-    fn recompute_excludes_observer_light_from_subnet_share_limits() {
-        let peers = [
-            PeerId::random(),
-            PeerId::random(),
-            PeerId::random(),
-            PeerId::random(),
-        ];
-        let discovery_sources = vec![
-            PeerDiscoverySource::StaticBootstrap,
-            PeerDiscoverySource::Dht,
-        ];
-        let discovered = HashMap::from([
-            (peers[0], sample_record(peers[0], discovery_sources.clone())),
-            (peers[1], sample_record(peers[1], discovery_sources.clone())),
-            (peers[2], sample_record(peers[2], discovery_sources.clone())),
-            (
-                peers[3],
-                sample_record_with_role(
-                    peers[3],
-                    PeerNodeRole::ObserverLight,
-                    discovery_sources.clone(),
-                ),
-            ),
-        ]);
-        let active = HashMap::from([
-            (
-                peers[0],
-                transport_path(
-                    peers[0],
-                    "/ip4/192.168.10.1/udp/4101/quic-v1",
-                    TransportPathKind::Direct,
-                ),
-            ),
-            (
-                peers[1],
-                transport_path(
-                    peers[1],
-                    "/ip4/192.168.10.2/udp/4102/quic-v1",
-                    TransportPathKind::Direct,
-                ),
-            ),
-            (
-                peers[2],
-                transport_path(
-                    peers[2],
-                    "/ip4/10.20.30.40/udp/4103/quic-v1",
-                    TransportPathKind::Direct,
-                ),
-            ),
-            (
-                peers[3],
-                transport_path(
-                    peers[3],
-                    "/ip4/192.168.10.89/udp/4104/quic-v1",
-                    TransportPathKind::Direct,
-                ),
-            ),
-        ]);
-
-        let healths =
-            recompute_peer_manager_healths(&discovered, &active, &PeerManagerPolicy::default());
-        for peer in peers {
-            assert_eq!(healths[&peer].status, PeerManagerHealthStatus::Active);
-            assert!(
-                !healths[&peer].issues.iter().any(|issue| matches!(
-                    issue,
-                    PeerManagerHealthIssue::Ipv4SubnetConcentration { .. }
-                )),
-                "observer-light peers should not push the provider active set over subnet share limits: {:?}",
-                healths[&peer].issues
-            );
-        }
-    }
-
-    #[test]
-    fn recompute_still_blocks_provider_subnet_concentration() {
-        let peers = [
-            PeerId::random(),
-            PeerId::random(),
-            PeerId::random(),
-            PeerId::random(),
-        ];
-        let discovery_sources = vec![
-            PeerDiscoverySource::StaticBootstrap,
-            PeerDiscoverySource::Dht,
-        ];
-        let discovered =
-            HashMap::from(peers.map(|peer| (peer, sample_record(peer, discovery_sources.clone()))));
-        let active = HashMap::from([
-            (
-                peers[0],
-                transport_path(
-                    peers[0],
-                    "/ip4/192.168.10.1/udp/4101/quic-v1",
-                    TransportPathKind::Direct,
-                ),
-            ),
-            (
-                peers[1],
-                transport_path(
-                    peers[1],
-                    "/ip4/192.168.10.2/udp/4102/quic-v1",
-                    TransportPathKind::Direct,
-                ),
-            ),
-            (
-                peers[2],
-                transport_path(
-                    peers[2],
-                    "/ip4/10.20.30.40/udp/4103/quic-v1",
-                    TransportPathKind::Direct,
-                ),
-            ),
-            (
-                peers[3],
-                transport_path(
-                    peers[3],
-                    "/ip4/192.168.10.89/udp/4104/quic-v1",
-                    TransportPathKind::Direct,
-                ),
-            ),
-        ]);
-
-        let healths =
-            recompute_peer_manager_healths(&discovered, &active, &PeerManagerPolicy::default());
-        for peer in [peers[0], peers[1], peers[3]] {
-            assert_eq!(healths[&peer].status, PeerManagerHealthStatus::Blocked);
-            assert!(healths[&peer].issues.iter().any(|issue| matches!(
-                issue,
-                PeerManagerHealthIssue::Ipv4SubnetConcentration {
-                    peers_in_bucket: 3,
-                    active_peer_count: 4,
-                    ..
-                }
-            )));
-        }
-        assert_eq!(healths[&peers[2]].status, PeerManagerHealthStatus::Active);
     }
 
     #[test]
