+++ crates/oasis7_net/src/libp2p_net/tests.rs
@@ -22,6 +22,7 @@ mod admissible_request_peers_tests;
 mod discovery_dial_tests;
 mod discovery_peer_record_tests;
 mod peer_health_refresh_tests;
+mod peer_manager_observer_tests;
 mod peer_record_tests;
 mod request_peer_filter_tests;
 mod subscribe_ack_tests;
